+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Added
 * Add support for blacklisting / whitelisting hosts to the HTTP runner by adding new
   ``url_hosts_blacklist`` and ``url_hosts_whitelist`` runner attribute. (new feature)
   #4757
+* Add ``user`` parameter to ``re_run`` method of st2client. #4785
 
 Changed
 ~~~~~~~